+++ tests/unit/test_utils.py
@@ -558,7 +558,7 @@ def test_repack_model_from_file_to_file(tmp):
     model_tar_path = os.path.join(tmp, "model.tar.gz")
     sagemaker.utils.create_tar_file([os.path.join(tmp, "model")], model_tar_path)
 
-    sagemaker_session = MagicMock()
+    sagemaker_session = MagicMock(settings=SessionSettings())
 
     file_mode_path = "file://%s" % model_tar_path
     destination_path = "file://%s" % os.path.join(tmp, "repacked-model.tar.gz")
@@ -608,7 +608,7 @@ def test_repack_model_from_file_to_folder(tmp):
         [],
         file_mode_path,
         "file://%s/repacked-model.tar.gz" % tmp,
-        MagicMock(),
+        MagicMock(settings=SessionSettings()),
     )
 
     assert list_tar_files("file://%s/repacked-model.tar.gz" % tmp, tmp) == {
@@ -679,7 +679,7 @@ def test_repack_model_with_same_inference_file_name(tmp, fake_s3):
 class FakeS3(object):
     def __init__(self, tmp):
         self.tmp = tmp
-        self.sagemaker_session = MagicMock()
+        self.sagemaker_session = MagicMock(settings=SessionSettings())
         self.location_map = {}
         self.current_bucket = None
         self.object_mock = MagicMock()
@@ -755,6 +755,10 @@ def test_sts_regional_endpoint():
 
     endpoint = sagemaker.utils.sts_regional_endpoint("us-iso-east-1")
     assert endpoint == "https://sts.us-iso-east-1.c2s.ic.gov"
+    assert botocore.utils.is_valid_endpoint_url(endpoint)
+
+    endpoint = sagemaker.utils.sts_regional_endpoint("us-isob-east-1")
+    assert endpoint == "https://sts.us-isob-east-1.sc2s.sgov.gov"
     assert botocore.utils.is_valid_endpoint_url(endpoint)